Update: change in Id type
Change Integer type to Long type for better consistency

+++ src/main/java/com/backend/agrosensor/agrosensorbackend/entity/base/AbstractUser.java
@@ -11,7 +11,7 @@ import lombok.Data;
 public abstract class AbstractUser {
 
     @Id
-    private Integer cc;
+    private Long cc;
     private String name;
     private String lastname;
